requirements.txt
joblib==0.17.0
numpy==1.19.4
Pillow==8.0.1
scikit-learn==0.23.2
scipy==1.5.4
sklearn==0.0
threadpoolctl==2.1.0
